Update all block related tests
